+++ MIP/src/Numeric/Optimization/MIP/Solver/CBC.hs
@@ -17,10 +17,12 @@ module Numeric.Optimization.MIP.Solver.CBC
   , cbc
   ) where
 
+import Control.Monad
 import Data.Default.Class
 import qualified Data.Text.Lazy.IO as TLIO
+import System.Directory
 import System.Exit
-import System.IO
+import System.FilePath
 import System.IO.Temp
 import qualified Numeric.Optimization.MIP.Base as MIP
 import qualified Numeric.Optimization.MIP.LPFile as LPFile
@@ -49,35 +51,36 @@ instance IsSolver CBC IO where
     case LPFile.render def prob{ MIP.objectiveFunction = obj' } of
       Left err -> ioError $ userError err
       Right lp -> do
-        withSystemTempFile "cbc.lp" $ \fname1 h1 -> do
-          TLIO.hPutStr h1 lp
-          hClose h1
-          withSystemTempFile "cbc.sol" $ \fname2 h2 -> do
-            hClose h2
-            let args = cbcArgs solver
-                    ++ [fname1]
-                    ++ (case solveTimeLimit opt of
-                          Nothing -> []
-                          Just sec -> ["sec", show sec])
-                    ++ (case solveTol opt of
-                          Nothing -> []
-                          Just tol ->
-                            [ "integerTolerance", show (MIP.integralityTol tol)
-                            , "primalTolerance", show (MIP.feasibilityTol tol)
-                            , "dualTolerance", show (MIP.optimalityTol tol)
-                            ])
-                    ++ ["solve", "solu", fname2]
-                onGetLine = solveLogger opt
-                onGetErrorLine = solveErrorLogger opt
-            exitcode <- runProcessWithOutputCallback (cbcPath solver) args Nothing "" onGetLine onGetErrorLine
-            case exitcode of
-              ExitFailure n -> ioError $ userError $ "exit with " ++ show n
-              ExitSuccess -> do
-                sol <- CBCSol.readFile fname2
-                if isMax then
-                  return $ sol{ MIP.solObjectiveValue = fmap negate (MIP.solObjectiveValue sol) }
-                else
-                  return sol
+        withSystemTempDirectory "haskell-mip-cbc" $ \dir -> do
+          let fname1 = dir </> "cbc.lp"
+              fname2 = dir </> "cbc.sol"
+          TLIO.writeFile fname1 lp
+          let args = cbcArgs solver
+                  ++ [fname1]
+                  ++ (case solveTimeLimit opt of
+                        Nothing -> []
+                        Just sec -> ["sec", show sec])
+                  ++ (case solveTol opt of
+                        Nothing -> []
+                        Just tol ->
+                          [ "integerTolerance", show (MIP.integralityTol tol)
+                          , "primalTolerance", show (MIP.feasibilityTol tol)
+                          , "dualTolerance", show (MIP.optimalityTol tol)
+                          ])
+                  ++ ["solve", "solu", fname2]
+              onGetLine = solveLogger opt
+              onGetErrorLine = solveErrorLogger opt
+          exitcode <- runProcessWithOutputCallback (cbcPath solver) args Nothing "" onGetLine onGetErrorLine
+          case exitcode of
+            ExitFailure n -> ioError $ userError $ "exit with " ++ show n
+            ExitSuccess -> do
+              m <- doesFileExist fname2
+              unless m $ ioError $ userError "CBC returned exit code 0, but wrote no solution file. You may want to use the solveLogger or solveErrorLogger for more information"
+              sol <- CBCSol.readFile fname2
+              if isMax then
+                return $ sol{ MIP.solObjectiveValue = fmap negate (MIP.solObjectiveValue sol) }
+              else
+                return sol
     where
       obj = MIP.objectiveFunction prob
       isMax = MIP.objDir obj == MIP.OptMax