+++ MIP/ChangeLog.md
@@ -4,6 +4,7 @@
 
 * Add `Numeric.Optimization.MIP.Solution.MIPLIB` module for reading/writing MIPLIB solution format
 * Add `Domain` type
+* Rename `S1` and `S2` to `SOS1` and `SOS2` respectively. Old names are kept as pattern synonyms.
 
 ## 0.2.0.0 (2025-02-03)