+++ MIP/ChangeLog.md
@@ -4,6 +4,7 @@
 
 * Add `Numeric.Optimization.MIP.Solution.MIPLIB` module for reading/writing MIPLIB solution format
 * Add `Domain` type
+* Add utility functions: `isInDomain`, `isIntegral`, `isInBounds`, and `constrBounds`
 
 ## 0.2.0.0 (2025-02-03)