+++ MIP/ChangeLog.md
@@ -6,6 +6,7 @@
 * Add `Domain` type
 * Add utility functions: `isInDomain`, `isIntegral`, `isInBounds`, and `constrBounds`
 * Rename `S1` and `S2` to `SOS1` and `SOS2` respectively. Old names are kept as pattern synonyms.
+* Make `Tol` an instance of `Eq`, `Ord`, and `Show`
 
 ## 0.2.0.0 (2025-02-03)