+++ MIP/ChangeLog.md
@@ -8,6 +8,7 @@
 * Rename `S1` and `S2` to `SOS1` and `SOS2` respectively. Old names are kept as pattern synonyms.
 * Make `Tol` an instance of `Eq`, `Ord`, and `Show`
 * Add `Default` instance for `SOSConstraint`
+* Produce more user-friendly error messages when the CBC solver exited with 0 but the output file was not generated (thanks to @dpvanbalen)
 
 ## 0.2.0.0 (2025-02-03)